refactor: Event를 통해 테이블 단체 지정 해제 로직을 OrderTable과 분리
