--- src/main/java/kitchenpos/table/domain/OrderTablesValidator.java
@@ -1,6 +1,0 @@
-package kitchenpos.table.domain;
-
-public interface OrderTablesValidator {
-    
-    void validate(final OrderTables orderTables);
-}
