+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -3,15 +3,13 @@ package kitchenpos.tablegroup.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.table.application.OrderTablesMapper;
-import kitchenpos.table.application.dto.GroupOrderTableRequest;
-import kitchenpos.table.application.dto.TableGroupResult;
-import kitchenpos.table.application.dto.TableGroupingRequest;
-import kitchenpos.table.domain.OrderTables;
-import kitchenpos.table.domain.OrderTablesValidator;
+import kitchenpos.tablegroup.application.dto.GroupOrderTableRequest;
+import kitchenpos.tablegroup.application.dto.TableGroupResult;
+import kitchenpos.tablegroup.application.dto.TableGroupingRequest;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.tablegroup.domain.TableGroupingEvent;
+import kitchenpos.tablegroup.domain.TableUngroupEvent;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,19 +19,13 @@ import org.springframework.transaction.annotation.Transactional;
 public class TableGroupService {
 
     private final TableGroupRepository tableGroupRepository;
-    private final OrderTablesMapper orderTablesMapper;
-    private final OrderTablesValidator ordersStatusValidator;
     private final ApplicationEventPublisher eventPublisher;
 
     public TableGroupService(
             final TableGroupRepository tableGroupRepository,
-            final OrderTablesMapper orderTablesMapper,
-            final OrderTablesValidator ordersStatusValidator,
             final ApplicationEventPublisher eventPublisher
     ) {
         this.tableGroupRepository = tableGroupRepository;
-        this.orderTablesMapper = orderTablesMapper;
-        this.ordersStatusValidator = ordersStatusValidator;
         this.eventPublisher = eventPublisher;
     }
 
@@ -51,7 +43,8 @@ public class TableGroupService {
     }
 
     public void ungroup(final Long ungroupTableId) {
-        final OrderTables orderTables = orderTablesMapper.fromTable(ungroupTableId);
-        orderTables.ungroup(ordersStatusValidator);
+        final TableGroup tableGroup = tableGroupRepository.findById(ungroupTableId)
+                .orElseThrow(() -> new IllegalArgumentException("Table group does not exist."));
+        eventPublisher.publishEvent(new TableUngroupEvent(tableGroup.getId()));
     }
 }
