+++ src/main/java/kitchenpos/tablegroup/application/dto/GroupOrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.application.dto;
+package kitchenpos.tablegroup.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
