+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -6,10 +6,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import kitchenpos.generic.IntegrationTest;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.table.application.dto.GroupOrderTableRequest;
-import kitchenpos.table.application.dto.TableGroupResult;
-import kitchenpos.table.application.dto.TableGroupingRequest;
 import kitchenpos.table.domain.OrderTable;
+import kitchenpos.tablegroup.application.dto.GroupOrderTableRequest;
+import kitchenpos.tablegroup.application.dto.TableGroupResult;
+import kitchenpos.tablegroup.application.dto.TableGroupingRequest;
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
